+++ renewit/api/urls.py
@@ -1,10 +1,10 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from .views import ( UpcyclerViewSet, UpcyclerClothesRequestViewSet,
+from .views import ( UserViewSet, UpcyclerClothesRequestViewSet,
 )
 
 router = DefaultRouter()
-router.register(r'upcyclers', UpcyclerViewSet, basename="upcyclers")
+router.register(r'user', UserViewSet, basename="user")
 router.register(r'clothes-requests', UpcyclerClothesRequestViewSet, basename="clothes-requests")
